requirements.txt
gradio
google-generativeai
google-cloud-bigquery
python-dotenv
